Assertions and logging for FFL_RESULT_ERROR cases, misc. linting changes
* FFLCharModel.h, FFLModelType.h: Some casts to int to satisfy
ClangSharp..??
* FFLGlobal.cpp: Add stub for FFLGladLoadGL
* FFLShaderCallback.cpp, FFLInitCharModelCPUStep,
FFLiCharModelCreateParam.cpp, GetStoreData(), FFLiManager::Create,
FFLiResourceLoader.cpp, FFLiResourceHeader::GetResult(): Add logs and
asserts that will help track down FFL_RESULTs.
* Misc linting changes.
* FFLiCharModelCreateParam.cpp: Log all of the FFLCharModelDesc params
that failed verification.

+++ include/nn/ffl/FFLCharModel.h
@@ -11,9 +11,8 @@
 extern "C" {
 #endif
 
-#define FFL_CHAR_MODEL_SIZE sizeof(FFLiCharModel)
-
-
+#define FFL_CHAR_MODEL_SIZE (int)sizeof(FFLiCharModel)
+// ^^ Cast is because ClangSharp says Cannot implicitly convert type int to ulong
 
 /*
 #if INTPTR_MAX == INT64_MAX
